// frontend/src/api/models/adminTrackDto.ts
/**
 * Generated by orval v8.5.1 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { AlbumSummaryDto } from './albumSummaryDto';
import type { ArtistSummaryDto } from './artistSummaryDto';
import type { PlaybackLinkDto } from './playbackLinkDto';

export interface AdminTrackDto {
  id?: string;
  title?: string;
  isrc?: string;
  durationMs?: number;
  tempoBpm?: number;
  createdAt?: string;
  processingStatus?: string;
  isFlagged?: boolean;
  flagReason?: string;
  hasVocals?: boolean;
  musicGenre?: string;
  genreConfidence?: number;
  artists?: ArtistSummaryDto[];
  album?: AlbumSummaryDto;
  playbackLinks?: PlaybackLinkDto[];
  danceStyle?: string;
  subStyle?: string;
  tempoCategory?: string;
  confidence?: number;
}