// frontend/src/api/models/adminTrackPageResponse.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { AdminTrackDto } from './adminTrackDto';

export interface AdminTrackPageResponse {
  items?: AdminTrackDto[];
  total?: number;
  page?: number;
  size?: number;
  hasMore?: boolean;
}